seems to work, but:
if the viewer leaves during person video running, does
nothing and lets the person video finish

+++ show1.py
@@ -27,7 +27,7 @@ persons = glob.glob(VIDEO_PATH + "/P*")
 videofiles = zip(landscapes, persons)
 videofiles = itertools.cycle(videofiles)
 
-# here we will define so variables in the global scope which are used in the functions
+# here we will define the variables in the global scope which are used in functions
 person_started = None
 last_sensor_state = None
 person_process = None
@@ -39,6 +39,7 @@ state = STATE_START
 
 
 def start_video(filename):
+    print "starting", filename
     return subprocess.Popen([PLAYER, filename], stdin=subprocess.PIPE)
 
 def start_next_landscape():
@@ -52,7 +53,11 @@ def start_person():
     person_started = time.time()
 
 def kill_landscape():
-    landscape_process.kill()
+    global landscape_process
+    if landscape_process:
+        print "killing", landscape_process
+        landscape_process.kill()
+    landscape_process = None
     
 def error():
     print "Unexpected state: %s, event: %s" % (state, event)
@@ -65,7 +70,7 @@ while True:
     current_time = time.time()
     if person_started and current_time - person_started > START_TIMEOUT:
         event = EVENT_TIMEOUT
-    person_started = None
+        person_started = None
     new_sensor_state = sensor.state()
     if last_sensor_state != new_sensor_state:
         if new_sensor_state == True:
@@ -80,15 +85,15 @@ while True:
         event = EVENT_FINISHED
         landscape_process = None
     
-    #print "event: ", event
-    #print "state: ", state
-    if state == STATE_START:
+    if state == STATE_START and event != EVENT_INCOMING:
         start_next_landscape()
         state = STATE_LANDSCAPE_RUNNING
     if not event:
         continue
+    print "event: ", event
+    print "state: ", state
     # TODO: if person video is running, do not start next landscape video
-    elif state == STATE_LANDSCAPE_RUNNING:
+    if state == STATE_LANDSCAPE_RUNNING:
         if event == EVENT_FINISHED:
             start_next_landscape()
             state = STATE_LANDSCAPE_RUNNING
@@ -103,8 +108,9 @@ while True:
             error()
     elif state == STATE_PERSON_STARTING:
         if event == EVENT_FINISHED:
-            start_next_landscape()
-            state = STATE_LANDSCAPE_RUNNING
+            # person video was started just before the lanscape video finished
+            # don't do anything
+            pass
         elif event == EVENT_INCOMING:
             state = STATE_PERSON_STARTING
         elif event == EVENT_OUTGOING:
@@ -122,8 +128,9 @@ while True:
         elif event == EVENT_INCOMING:
             state = STATE_PERSON_RUNNING
         elif event == EVENT_OUTGOING:
-            start_next_landscape()
-            state = STATE_LANDSCAPE_RUNNING
+            # if the viewer leaves during person video running, do nothing and
+            # let the person video finish
+            pass
         elif event == EVENT_TIMEOUT:
             error()
         else:
